TESB-21304 NPE during importing of Route is fixed
At this moment we have NPE during invocation of "JavaCamelJobScriptsExportWSAction" from "BundleJavaProcessor".
The reason of NPE is   "repositoryNode" (from BundleJavaProcessor) is null because of Eclipse somehow refreshes TreeView after finishing of Workspace opertion only. And imported element can not be retrived from TreeVeiw before Import operation is finished. Current fix is just workaround for such situation. It will be necessary to review this code when propagation of Build Type will be fully implemented (we must avoid using BundleJavaProcessor for imported Routes).